backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic>=2.7.4,<3.0.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-multipart==0.0.6
pyjwt==2.8.0
langchain==0.3.25
langchain-core==0.3.58
langchain-openai==0.3.16
langchain-community==0.3.23
langchain-text-splitters==0.3.8
openai>=1.68.2
pypdf==4.0.1
faiss-cpu==1.7.4
python-jose==3.3.0
alembic==1.13.1
passlib==1.7.4
bcrypt==4.1.2
tenacity==8.2.3 